Clean code

+++ Kint-Sources-master/OpenNos.GameObject/Packets/ClientPackets/CRegPacket.cs
@@ -24,7 +24,7 @@ namespace OpenNos.GameObject
         [PacketIndex(6)]
         public int IsPackage { get; set; }
         [PacketIndex(7)]
-        public byte Amount { get; set; }
+        public short Amount { get; set; }
         [PacketIndex(8)]
         public long Price { get; set; }
         [PacketIndex(9)]
